DS-1804: move common/[dal/indexing] to datasources
