--- common/indexing/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "common-indexing"
-version = "0.1.0"
-authors = ["Datafuse Authors <opensource@datafuselabs.com>"]
-license = "Apache-2.0"
-publish = false
-edition = "2021"
-
-[dependencies] # In alphabetical order
-# Workspace dependencies
-common-datablocks = {path = "../datablocks"}
-common-datavalues = {path = "../datavalues"}
-common-exception = {path = "../exception"}
-common-planners = {path = "../planners"}
-
-# Crates.io dependencies
-serde = { version = "1.0", features = ["derive"] }
-
-[dev-dependencies]
-pretty_assertions = "0.7"
