--- common/indexing/src/index_partition.rs
@@ -1,36 +1,0 @@
-// Copyright 2021 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-use common_datavalues::DataValue;
-use common_exception::Result;
-use common_planners::Expression;
-
-/// Min and Max index.
-#[derive(Clone, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
-pub struct PartitionIndex {}
-
-impl PartitionIndex {
-    pub fn typ(&self) -> &str {
-        "partition"
-    }
-
-    /// Apply the expr against the partition value, and get the result:
-    /// true: need
-    /// false: skip
-    pub fn apply_index(_partition_value: DataValue, _expr: &Expression) -> Result<bool> {
-        // TODO(bohu): expression apply.
-        Ok(true)
-    }
-}
