+++ query/src/datasources/dal/blob_accessor.rs
