+++ query/src/sessions/context.rs
@@ -18,10 +18,6 @@ use std::sync::atomic::Ordering;
 use std::sync::atomic::Ordering::Acquire;
 use std::sync::Arc;
 
-use common_dal::DataAccessor;
-use common_dal::Local;
-use common_dal::StorageScheme;
-use common_dal::S3;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_infallible::RwLock;
@@ -43,6 +39,10 @@ use crate::catalogs::TableFunctionMeta;
 use crate::catalogs::TableMeta;
 use crate::clusters::ClusterRef;
 use crate::configs::Config;
+use crate::datasources::dal::DataAccessor;
+use crate::datasources::dal::Local;
+use crate::datasources::dal::StorageScheme;
+use crate::datasources::dal::S3;
 use crate::sessions::context_shared::DatafuseQueryContextShared;
 use crate::sessions::SessionManagerRef;
 use crate::sessions::Settings;
